+++ src/SonarScanner.MSBuild.PreProcessor/AnalysisConfigProcessing/Processors/PropertyMappingProcessor.cs
@@ -20,6 +20,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.IO;
 using SonarScanner.MSBuild.Common;
 
 namespace SonarScanner.MSBuild.PreProcessor.AnalysisConfigProcessing.Processors;
@@ -32,13 +33,13 @@ public class PropertyMappingProcessor(ProcessedArgs localSettings, IDictionary<s
 {
     private static readonly IDictionary<string, string> MappedId = new Dictionary<string, string>
     {
-        { "sonar.scanner.truststorePath", "javax.net.ssl.trustStore" },
-        { "sonar.scanner.truststorePassword", "javax.net.ssl.trustStorePassword" }
+        { SonarProperties.TruststorePath, "javax.net.ssl.trustStore" },
+        { SonarProperties.TruststorePassword, "javax.net.ssl.trustStorePassword" }
     };
 
     private static readonly IDictionary<string, Func<string, string>> MappedValue = new Dictionary<string, Func<string, string>>
     {
-        { "sonar.scanner.truststorePath", ConvertToJavaPath },
+        { SonarProperties.TruststorePath, ConvertToJavaPath },
     };
 
     public override void Update(AnalysisConfig config)
@@ -57,5 +58,5 @@ public class PropertyMappingProcessor(ProcessedArgs localSettings, IDictionary<s
     }
 
     private static string ConvertToJavaPath(string path) =>
-        path?.Replace('\\', '/');
+        path?.Replace(Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar);
 }
